* Fixed bug in Widcomm Bluetooth discovery * Better return value checks when allocating data. * Check max size on metadata header when retrieving * Lots of other changes for robustness

+++ src/hagglekernel/ConnectivityEthernet.h
@@ -27,7 +27,7 @@
 
 struct haggle_beacon {
         u_int32_t seqno;
-	u_int8_t interval; // The beacon interval used by the other node (in seconds)
+	u_int32_t interval; // The beacon interval used by the other node (in seconds)
         char mac[6];
         char pad[2];
 };
